[FIX] Handling segfault when plane in corner

+++ src/load_entities.c
@@ -92,10 +92,10 @@ int add_to_plane(aircraft_t *aircraft, char **array, int plane_added)
     aircraft[plane_added].delay = (float)my_getnbr(array[6]);
     aircraft[plane_added].arrived = FALSE;
     aircraft[plane_added].destroyed = FALSE;
-    if (aircraft[plane_added].x_departure > (float)WIDTH ||
-        aircraft[plane_added].x_arrival > (float)WIDTH ||
-        aircraft[plane_added].y_departure > (float)HEIGHT ||
-        aircraft[plane_added].y_arrival > (float)HEIGHT)
+    if (aircraft[plane_added].x_departure > WIDTH ||
+        aircraft[plane_added].x_arrival > WIDTH ||
+        aircraft[plane_added].y_departure > HEIGHT ||
+        aircraft[plane_added].y_arrival > HEIGHT)
         return FAILURE;
     return SUCCESS;
 }
